% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best.friends-package.r
\docType{package}
\name{best.friends-package}
\alias{best.friends-package}
\title{best.friends: A package that describe whether a collection is friendly for a tag of not}
\description{
We have a set T of tags a set C of collections. Clouds pay attention to tags.
The attention that each cloud pays to each tag is represented by a real value.
The attention  can represent any type of relation measure, e.g. fuzzy membership.
The attention values is a \eqn{|T| \times |C|} matrix \eqn{A}.
}
\details{
For each tag, we want to identify the collection(s) that specifically prefer(s) the tag.
We say that such a collection is a friend (or the best friend if it is the only) for the tag.
The simplest example: imagine that only one collection pays attention to our tag.

See our arxiv paper for details.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/favorov/best-friends}
  \item Report bugs at \url{https://github.com/favorov/best-friends/issues}
}

}
\author{
\strong{Maintainer}: Alexander Favorov \email{favorov@sensi.org} (\href{https://orcid.org/0000-0003-2595-1091}{ORCID})

Authors:
\itemize{
  \item Alexandra Suvorikova \email{a.suvorikova@gmail.com} [contributor]
  \item Alexey Kroshnin \email{kroshnin@wias-berlin.de} [contributor]
  \item Dmitrijs Lvovs \email{dmitrijs.lvovs@gmail.com} [contributor]
  \item Vera Mukhina \email{atgc@vigg.ru} [contributor]
  \item Vasily Ramensky \email{ramensky@gmail.com} [contributor]
  \item Andrey A. Mironov \email{mironov@bioinf.fbb.msu.ru} [contributor]
}

}
\keyword{internal}
