% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best.friends.bic.r
\name{best.friends.bic}
\alias{best.friends.bic}
\title{best.friends.bic}
\usage{
best.friends.bic(attention = NULL, prior.to.have.friends = -1, best.no = 1)
}
\arguments{
\item{attention}{original attention matrix}

\item{prior.to.have.friends}{The prior for a tag is important enough to have friendly collections.}

\item{best.no}{The maximal number of friends for a tag, the default is \code{1},
i.e. the best friend. The string "all" means "all friends".
The value $n$ means that we filter out a tag from the results is it has more
than $n$ friends.}
}
\value{
A data.frame, rows are pairs of tags and collections that are markers and best friends
friends.
}
\description{
Find Tags that are best friends to Collections
}
\examples{
attention <- matrix(c(10,6,7,8,9,
                9,10,6,7,8,
                8,9,10,6,7,
                7,8,9,10,6,
                6,7,8,9,10,
                20,0,0,0,0), 
                nrow=6, ncol=5, byrow=TRUE)
res <- best.friends.bic(attention, prior.to.have.friends=0.5)
}
