% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag.int.ranks.r
\name{tag.int.ranks}
\alias{tag.int.ranks}
\title{tag.int.ranks}
\usage{
tag.int.ranks(
  attention = NULL,
  distance_like = FALSE,
  neglect_diagonal = FALSE
)
}
\arguments{
\item{attention}{original attention matrix}

\item{distance_like}{if \code{TRUE}, ranks are ascending (the smaller is the better). Default is \code{FALSE}.}

\item{neglect_diagonal}{if square matrix, and the diagonal does not make sense, we \code{NA} it and our ranks are in \eqn{[1 .. |T|-1]}. Default is \code{FALSE}.}
}
\value{
a \eqn{|T| x |C|} matrix of integer ranks in \eqn{[1 .. |T|]} (or, \eqn{[1 .. |T|-1]} if neglect_diagonal).
}
\description{
Given the (\code{attention}) \eqn{|T| \times |C|} matrix, convert it to matrix of descending ranks of tags(rows) in collections (columns).
#' See \link{best.friends} documentation for details.
}
\examples{
genes<-10
regulation<-matrix(
    c(0.2, 0.2, 0.2, 0.2, 0.25, rep(0.2,genes-5),
        rep(1, genes),
        rep(1, genes),
        rep(1, genes),
        rep(1, genes),
        rep(1, genes),
        rep(1, genes),
        rep(1, genes),
        rep(1, genes),
        rep(1, genes)
     ),
    ncol=10,byrow=FALSE
)
gene.names<-LETTERS[seq( from = 1, to = genes )]
TF.names<-c('TF1','TF2','TF3','TF4','TF5','TF6','TF7','TF8','TF9','TF10')
rownames(regulation)<-gene.names
colnames(regulation)<-TF.names
TF.ranks<-tag.int.ranks(regulation)
}
