% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag.ranks.r
\name{tag.ranks}
\alias{tag.ranks}
\title{tag.ranks}
\usage{
tag.ranks(attention, distance_like = FALSE, neglect_diagonal = FALSE)
}
\arguments{
\item{attention}{is the \eqn{|T| \times |C|} matrix of the relations between tags and the clouds.}

\item{distance_like}{the default is \code{FALSE} and it shows that the relation values are
not like distance, i.e. the better relation is shown by the larger value;
if the relation is, on the contrary, distance-like, and 0 is the best, the value is \code{TRUE}.}

\item{neglect_diagonal}{in the case of square attention matrix, the diagonal sometimes is either
uninformative or it carries some specific values. In each of these cases,
the diagonal elements are excluded from the ranking and from the statistics
by setting this parameter TRUE. The default is FALSE.}
}
\value{
The matrix of ranks of attention to tags from collections. The size is the same as that of the attention matrix.
collection.
}
\description{
tag.ranks
}
\examples{
genes<-10
regulation<-matrix(
  c(0.2, 0.2, 0.2, 0.2, 0.25, rep(0.2,genes-5),
    rep(1, genes),
    rep(1, genes),
    rep(1, genes),
    rep(1, genes),
    rep(1, genes),
    rep(1, genes),
    rep(1, genes),
    rep(1, genes),
    rep(1, genes)
  ),
  ncol=10,byrow=FALSE
)
gene.names<-LETTERS[seq( from = 1, to = genes )]
TF.names<-c('TF1','TF2','TF3','TF4','TF5','TF6','TF7','TF8','TF9','TF10')
rownames(regulation)<-gene.names
colnames(regulation)<-TF.names
ranks<-tag.ranks(regulation)
bestfriends<-best.friends.test(regulation)
}
