% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best.friends-package.r
\docType{package}
\name{best.friends-package}
\alias{best.friends}
\alias{best.friends-package}
\title{best.friends: We look for collections that are friendly for the tags}
\description{
We have tags (rows) and collections (columns) and each tag-collection pair is described by the the value of the attention (real) the collection pays to the tag. Given the attention matrix, we want to identify whether a tag seperates all the collections by the relevance of the tag for the collections. The collection, for which the tag is relevant, is the tag's friend. The tag is the collection's marker.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/favorov/best-friends}
}

}
\author{
\strong{Maintainer}: Alexander Favorov \email{favorov@sensi.org} (\href{https://orcid.org/0000-0003-2595-1091}{ORCID})

Authors:
\itemize{
  \item Alexandra Suvorikova \email{a.suvorikova@gmail.com} [contributor]
  \item Dmitrijs Lvovs \email{dmitrijs.lvovs@gmail.com} [contributor]
  \item Vera Mukhina \email{atgc@vigg.ru} [contributor]
  \item Vasily Ramensky \email{ramensky@gmail.com} [contributor]
  \item Andrey A. Mironov \email{mironov@bioinf.fbb.msu.ru} [contributor]
}

}
\keyword{internal}
