% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best.friends.r
\name{best.friends}
\alias{best.friends}
\title{best.friends}
\usage{
best.friends(
  attention = NULL,
  threshold = 0.05,
  p.adjust.method = "BH",
  best.no = 1
)
}
\arguments{
\item{attention}{original attention matrix}

\item{threshold}{The adjusted p-value threshold for KS test for
non-uniformity of ranks.}

\item{p.adjust.method}{Multiple testing correction method, see \link[stats]{p.adjust}.}

\item{best.no}{The maximum number of friends for a tag, the default is \code{1},
i.e. the best friend. The string "all" means "all friends".}
}
\value{
A data those of tags and collections that are markers and best friends
friends.
}
\description{
Find Tags that are best friends to Collections
}
\examples{
attention <- matrix(c(10,6,7,8,9,
                9,10,6,7,8,
                8,9,10,6,7,
                7,8,9,10,6,
                6,7,8,9,10,
                20,0,0,0,0), nrow=6, ncol=5)
res <- best.friends(attention, threshold = 1)
}
