% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best.friends.r
\name{best.friends}
\alias{best.friends}
\title{best.friends}
\usage{
best.friends(mat, threshold = 0.05, p.adjust.method = "BH", best.no = 1)
}
\arguments{
\item{mat}{A matrix tags x collections of attention that a collection pays to tag.}

\item{threshold}{The adjusted p-value threshold for KS test for
non-uniformity of ranks.}

\item{p.adjust.method}{Multiple testing correction method, see \link[stast]{p.adjust}.}

\item{best.no}{The maximum number of friends for a tag, the default is \code{1},
i.e. the best friend.}
}
\value{
A data those of tags and collections that are markers and best friends
friends.
}
\description{
Find Tags that are best friends to Collections
}
\examples{
WIP
}
