% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best.friends.test.r
\name{best.friends.test}
\alias{best.friends.test}
\title{best.friends.test}
\usage{
best.friends.test(
  attention = NULL,
  ranks.of.tags = NULL,
  distance_like = FALSE,
  neglect_diagonal = FALSE
)
}
\arguments{
\item{attention}{is the tags*collections matrix of the relations between tags and the clouds}

\item{ranks.of.tags}{the value of \link{tag.ranks} call.
\code{best.friends.test(attenion)} and \code{best.friends.test(ranks.of.tags=tag.ranks(attenion)}
return the same, as well as
\code{friends.test(attenion)} and \code{friends.test(tag.ranks=tag.ranks(attenion)}.
The goal is to calculate tag.ranks only once if multiple calls of friends tests
happen for the same attention matrix.}

\item{distance_like}{the default is \code{FALSE} and it shows that the relation values are
not like distance, i.e. the better relation is shown by the larger value;
if the relation is, on the contrary, distance-like, and 0 is the best, the value is \code{TRUE}.}

\item{neglect_diagonal}{in the case of square attention matrix, the diagonal sometimes is either
uninformative or it carries some specific values. In each of these cases,
the diagonal elements are excluded from the ranking and from the statistics
by setting this parameter TRUE. The default is FALSE.}
}
\value{
\code{data.frame} with 5 columns: tag index,
the index of the collection that is a putative best friend of the tag,
uncorrected p-value for the pair,
tag,
collection.
The small (after multiple hypothesis correction we are to do) p-value
indicates that the collection is really the best friend of the tag.
}
\description{
Given the attention matrix for each tag, identifies the most friendly collection
and assesses the corresponding p-value.
}
\details{
If p-value is low and the null is rejected, the collection is the best friend for the tag
and the tag is the marker of the collection.

See \link{best.friends} documentation for details.
}
\examples{
genes<-10
regulation<-matrix(
  c(0.2, 0.2, 0.2, 0.2, 0.25, rep(0.2,genes-5),
    rep(1, genes),
    rep(1, genes),
    rep(1, genes),
    rep(1, genes),
    rep(1, genes),
    rep(1, genes),
    rep(1, genes),
    rep(1, genes),
    rep(1, genes)
  ),
  ncol=10,byrow=FALSE
)
gene.names<-LETTERS[seq( from = 1, to = genes )]
TF.names<-c('TF1','TF2','TF3','TF4','TF5','TF6','TF7','TF8','TF9','TF10')
rownames(regulation)<-gene.names
colnames(regulation)<-TF.names
bestfriends<-best.friends.test(regulation)
}
