% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best.step.fit.bic.r
\name{best.step.fit.bic}
\alias{best.step.fit.bic}
\title{best.step.fit.bic}
\usage{
best.step.fit.bic(ranks, tags.no, prior.to.have.friends)
}
\arguments{
\item{ranks}{vector of ranks of a tag in different collections}

\item{tags.no}{number of tags, i.e. maximal rank}

\item{prior.to.have.friends}{The prior for a tag is important enough to have friendly collections.}
}
\value{
a list of four values: \cr
\code{step.models} is return from \link{step.fit.ln.likelihoods} call the function start with
\code{best.step.rank} is the rank value that makes the best step;
it is not obligatory one on the \code{ranks} value.\cr
\code{collections.on.left} is the vector of the collections on the left of the best step
(including the step value). They are friends of the tag.\cr
\code{collections.on.right} is vector of those on the right \cr
\code{population.on.left} is how many ranks are on left of split; they are friends! \cr
if non-step uniform model wins and there are no friends,\cr
then \code{best.step.rank==tags.no}, \code{population.on.left==0},\cr
all collections are listed in \code{collections.on.right} and \code{collections.on.left} is empty
}
\description{
finds the ML-best step model and
compares the posteriors of the found best step
and non-step uniform model.
}
\details{
See \link{best.friends} documentation for details.
}
\examples{
example(tag.int.ranks)
steps<-best.step.fit.bic(TF.ranks[42,],genes.no,0.5)
}
