% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best.step.fit.r
\name{best.step.fit}
\alias{best.step.fit}
\title{best.step.fit}
\usage{
best.step.fit(ranks, tags.no)
}
\arguments{
\item{ranks}{vector of ranks of a tag in different collections}

\item{tags.no}{number of tags, i.e. maximal rank}
}
\value{
a list of four values: \cr
\code{step.models} is return from \link{step.fit.ln.likelihoods} call the function start with
\code{best.step.rank} is the rank value that makes the best step;
it is not obligatory one on the \code{ranks} value.\cr
\code{collections.on.left} is the vector of the collections on the left of the best step
(including the step value). They are friends of the tag.\cr
\code{collections.on.right} is vector of those on the right \cr
\code{population.on.left} is how many ranks are on left of split; they are friends! \cr
}
\description{
finds the ML-best step model
}
\details{
See \link{best.friends} documentation for details.
}
\examples{
example(tag.int.ranks)
steps<-step.fit.ln.likelihoods(TF.ranks[42,],genes.no)
}
