% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friends.test.r
\name{friends.test}
\alias{friends.test}
\title{friends.test}
\usage{
friends.test(
  attention = NULL,
  ranks.of.tags = NULL,
  distance_like = FALSE,
  friends.number = -1,
  neglect_diagonal = FALSE
)
}
\arguments{
\item{attention}{is the tags*collections matrix of the relations between tags and the clouds}

\item{ranks.of.tags}{the value of \link{tag.ranks} call. \code{best.friends.test(attenion)} and \code{best.friends.test(ranks.of.tags=tag.ranks(attenion)} return the same, as well as \code{friends.test(attenion)} and \code{friends.test(tag.ranks=tag.ranks(attenion)} The goal is to calculate tag.ranks only once if multiple calls of friends tests happen for the same attention matrix.}

\item{distance_like}{the default is \code{FALSE} and it shows that the relation values are not like distance,
i.e. the better relation is shown by the lagrer value; if the relation is, on the contrary, distance-like,
and 0 is the best, the value is \code{TRUE}.}

\item{friends.number}{number of entities we consider for each tags; the default -1 means all;
if friends.number is 1, the call does essentially the same as the best.friends.test call}

\item{neglect_diagonal}{in the case of square attention matrix, the diagonal sometimes is either uninformative or it carries some specific values. In each of these cases,
the diagonal elements are excluded from the ranking and from the statistics by setting this parameter TRUE. The default is FALSE.}
}
\value{
a list with 2 matrices, first is \eqn{|T| x |C|} dimensions (as the \code{attention}), the second is \eqn{|T| x (|C|-1)}.
if \code{friends.number} is some \eqn{F:1<=F<|C|-1}, the dimensions are \eqn{|T| x (F+1)} and \eqn{|T| x F}
\code{collections} for each tag, the the collections are ranked by the importance of the tag, best friend first;
\code{pvals} contains p-values for the corresponding split of the \code{collections} row to friends and others.
}
\description{
friends.test
}
\examples{
genes<-10
regulation<-matrix(
    c(0.2, 0.2, 0.2, 0.2, 0.25, rep(0.2,genes-5),
        rep(1, genes),
        rep(1, genes),
        rep(1, genes),
        rep(1, genes),
        rep(1, genes),
        rep(1, genes),
        rep(1, genes),
        rep(1, genes),
        rep(1, genes)
     ),
    ncol=10,byrow=FALSE
)
gene.names<-LETTERS[seq( from = 1, to = genes )]
TF.names<-c('TF1','TF2','TF3','TF4','TF5','TF6','TF7','TF8','TF9','TF10')
rownames(regulation)<-gene.names
colnames(regulation)<-TF.names
friends<-friends.test(regulation)
}
