% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.fit.r, R/step.ln.likelihoods.r
\name{step.ln.likelihoods}
\alias{step.ln.likelihoods}
\title{fit possible bi-uniform step models for a set of ranks
#' See \link{best.friends} documentation for details.}
\usage{
step.ln.likelihoods(ranks, tags.no)

step.ln.likelihoods(ranks, tags.no)
}
\arguments{
\item{ranks}{vector of ranks of a tag in different collections}

\item{tags.no}{number of tags, i.e. maximal rank}
}
\value{
the ln of the likelihood of the ranks for each step position in \eqn{1 .. max(ranks)+1}; the last is for uniform, non-step case
example(tag.int.ranks)
steps<-step.ln.likelihoods(TF.ranks\link{42,},genes.no)

list of three values: \code{ln.likelihoods} contains the ln of the likelihood of the ranks for each split (step) position into (this or less) and (greater than this) for \eqn{1 .. tags.no-1} the last (\eqn{tag.np)} element is for is for uniform, non-step case; \code{k1.by.l1} contains k_1 (number of values on the left of the step) for each split; \code{col.order} is the order of ranks
}
\description{
fit possible bi-uniform step models for a set of ranks
See \link{best.friends} documentation for details.
}
\examples{
example(tag.int.ranks)
steps<-step.ln.likelihoods(TF.ranks[42,],genes.no)
}
